--- a/cmd/factory_test.go
+++ b/cmd/factory_test.go
@@ -10,8 +10,8 @@ import (
 	"github.com/mitchellh/go-homedir"
 	"github.com/qri-io/ioes"
 	"github.com/qri-io/qri/config"
-	cfgtest "github.com/qri-io/qri/config/test"
 	"github.com/qri-io/qri/lib"
+	libtest "github.com/qri-io/qri/lib/test"
 	"github.com/qri-io/qri/p2p"
 	"github.com/qri-io/qri/repo"
 	"github.com/qri-io/qri/repo/gen"
@@ -56,7 +56,7 @@ func NewTestFactory(c *regclient.Client) (tf TestFactory, err error) {
 		IOStreams:   ioes.NewDiscardIOStreams(),
 		qriRepoPath: "",
 		ipfsFsPath:  "",
-		generator:   NewTestCrypto(),
+		generator:   libtest.NewTestCrypto(),
 
 		repo:   repo,
 		rpc:    nil,
@@ -192,25 +192,4 @@ func TestEnvPathFactory(t *testing.T) {
 		}
 
 	}
-}
-
-func NewTestCrypto() gen.CryptoGenerator {
-	return &testCryptoGenerator{}
-}
-
-type testCryptoGenerator struct {
-	count int
-}
-
-// TODO: Add a method to this that performs a version of `cafs/ipfs/init` or InitIPFS in
-// `actions/setup` that is more appropriate for tests.
-
-func (g *testCryptoGenerator) GeneratePrivateKeyAndPeerID() (string, string) {
-	info := cfgtest.GetTestPeerInfo(g.count)
-	g.count++
-	return info.EncodedPrivKey, info.EncodedPeerID
-}
-
-func (g *testCryptoGenerator) GenerateNickname(peerID string) string {
-	return "testnick"
 }