… lib/test/testdata/empty_ipfs_repo.zip