The last major speed problem in the unit tests was many cmd functions creating an empty IPFS repo. Now, tests using NewTestCrypto instead will extract a zip file that has an empty IPFS repo, avoiding the need to perform any primality testing.

Before
run 0
```
ok  	github.com/qri-io/qri/actions      3.158s
ok  	github.com/qri-io/qri/api          3.125s
ok  	github.com/qri-io/qri/cmd         22.842s
ok  	github.com/qri-io/qri/config       0.145s
ok  	github.com/qri-io/qri/lib          3.160s
ok  	github.com/qri-io/qri/p2p          3.389s
ok  	github.com/qri-io/qri/repo         0.116s
ok  	github.com/qri-io/qri/repo/fs      0.325s
ok  	github.com/qri-io/qri/repo/profile 0.105s
ok  	github.com/qri-io/qri/repo/test    0.193s
```
run 1
```
ok  	github.com/qri-io/qri/actions      2.609s
ok  	github.com/qri-io/qri/api          3.002s
ok  	github.com/qri-io/qri/cmd         20.146s
ok  	github.com/qri-io/qri/config       0.182s
ok  	github.com/qri-io/qri/lib          4.178s
ok  	github.com/qri-io/qri/p2p          3.378s
ok  	github.com/qri-io/qri/repo         0.120s
ok  	github.com/qri-io/qri/repo/fs      0.145s
ok  	github.com/qri-io/qri/repo/profile 0.092s
ok  	github.com/qri-io/qri/repo/test    0.257s
```

After
run 0
```
ok  	github.com/qri-io/qri/actions      3.053s
ok  	github.com/qri-io/qri/api          2.909s
ok  	github.com/qri-io/qri/cmd         16.784s
ok  	github.com/qri-io/qri/config       0.260s
ok  	github.com/qri-io/qri/lib          3.226s
ok  	github.com/qri-io/qri/p2p          3.230s
ok  	github.com/qri-io/qri/repo         0.235s
ok  	github.com/qri-io/qri/repo/fs      0.285s
ok  	github.com/qri-io/qri/repo/profile 0.099s
ok  	github.com/qri-io/qri/repo/test    0.272s
```
run 1
```
ok  	github.com/qri-io/qri/actions      2.812s
ok  	github.com/qri-io/qri/api          1.239s
ok  	github.com/qri-io/qri/cmd         17.399s
ok  	github.com/qri-io/qri/config       0.133s
ok  	github.com/qri-io/qri/lib          2.979s
ok  	github.com/qri-io/qri/p2p          3.202s
ok  	github.com/qri-io/qri/repo         0.144s
ok  	github.com/qri-io/qri/repo/fs      0.174s
ok  	github.com/qri-io/qri/repo/profile 0.111s
ok  	github.com/qri-io/qri/repo/test    0.311s
```